pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pytask_scheduler"
version = "2024.9.8"
requires-python = ">= 3.8"
dependencies = [
  "polars>=1.6.0",
  "pywin32>=303",
  "python-evtx>=0.7.4",
  "dataclasses>=0.6"
]
authors = [
    {name = "Jalen Merluza", email = "jalenmerluza12@gmail.com"}
]
description = "Windows Task Scheduler in Python!"
readme = "README.md"
license = {file = "LICENSE.txt"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python"
]

[project.urls]
Repository = "https://github.com/jmerluza/pytask_scheduler"